Se carga Persistencia avanzada Familia

+++ src/main/java/gc/_4/pr2/grupo2/service/jpa/FamiliaServiceImpl.java
@@ -11,27 +11,43 @@ import gc._4.pr2.grupo2.service.FamiliaService;
 
 @Service
 public class FamiliaServiceImpl implements FamiliaService {
+
+	// La variable repository es una instancia de la clase FamiliaRepository, que se utiliza para interactuar con la base de datos.
 	
 	@Autowired
-	private FamiliaRepository repository;
+    private FamiliaRepository repository;
+
+    @Override
+    public List<Familia> getFamilia() {
+        return repository.findAll();
+    }
+
+    @Override
+    public Familia getFamiliaById(Long id) {
+        return repository.findById(id).orElse(null);
+    }
+
+    @Override
+    public Familia saveFamilia(Familia familia) {
+        return repository.save(familia);
+    }
+
+   
+    @Override
+    public boolean deleteFamiliaById(Long id) {
+        if (repository.existsById(id)) {
+            repository.deleteById(id);
+            return true;
+        }
+        return false;
+    }
 
 	@Override
-	public List<Familia> getFamilia() {
-		return repository.findAll();
-	}
-	
-	@Override
-	public Familia getFamiliaById(Long id) {
-	    return repository.findById(id).orElse(null);
-	}
-	
-	@Override
-	public Familia saveFamilia(Familia familia) {
-		return repository.save(familia);
-	}
-
-	@Override
-	public void deteleFamiliaById(Long id) {
-		repository.deleteById(id);
+	public boolean existe(Long id) {
+		if(id == null) {
+			return false;
+		} else {
+			return repository.existsById(id);
+		}
 	}
 }
